rgw: fix distributed ratelimit build errors
Extract RGWRateLimitInfo into rgw_ratelimit_types.h so cls_rgw_ratelimit
does not pull rgw_common.h and RapidJSON. Remove invalid empty-string
default from rgw_ratelimit_rados_pool in rgw.yaml.in.
